feat(entities): add centralized entityNameMap and use it in viewer
Introduce a new entityNameMap module that exports a mapping of
entity keys to their React component implementations. Populate the map
with all existing entity components (blocks, props, raised-bed items,
etc.) to provide a single source of truth for available entities.

Update EntityGridViewer to derive its list of entities from
entityNameMap instead of requiring an external prop. This simplifies
the viewer API and ensures it always renders the canonical set of
entities. Adjust the dynamic wrapper in the debug app to stop passing
entityNames and just render the viewer directly.

Remove an incorrect re-export of entityNameMap from the package index
to avoid confusion and keep the mapping module as the explicit source.
