+++ packages/game/src/index.ts
@@ -1,4 +1,3 @@
-export { entityNameMap } from './entities/EntityFactory';
 export type { GameSceneProps } from './GameScene';
 export { GameSceneDynamic as GameScene } from './GameSceneDynamic';
 export { PlantEditor } from './generators/plant/editor/PlantEditor';
